requirements.txt
affine
click
fiona
geopandas
numpy
pandas
pyproj
pyyaml
rasterio
scikit-image
shapely
